+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased][unreleased]
 
 ### Added
-  - new proxy configuration dialog #5052
+- new proxy configuration dialog #5052
 
 ### Changed